% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats_ui.R
\name{create_dice_number_input}
\alias{create_dice_number_input}
\title{Creates dice number inputs}
\usage{
create_dice_number_input(
  input_id,
  dice_names = global$dice_names,
  default_dice_number = 0
)
}
\arguments{
\item{input_id}{prefix for inputIds}

\item{dice_names}{list of names for numericInput objects}

\item{default_dice_number}{initial value for numericInput objects}
}
\value{
list of shiny numericInput objects
}
\description{
Creates dice number inputs
}
