% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calculation.R
\name{create_stats_table}
\alias{create_stats_table}
\title{Creates dataframe with basic statistics for selected dice}
\usage{
create_stats_table(no_dice, rolls)
}
\value{
dataframe with basic statistics
}
\description{
Creates dataframe with basic statistics for selected dice
}
