% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_charts.R
\name{dice_histogram}
\alias{dice_histogram}
\title{Creates a histogram for given dice roll results}
\usage{
dice_histogram(results)
}
\arguments{
\item{results}{numeric vector containing roll results}
}
\value{
plotly histogram chart
}
\description{
Creates a histogram for given dice roll results
}
