% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calculation.R
\name{simulate_dice_roll}
\alias{simulate_dice_roll}
\title{Simulates dice rolls}
\usage{
simulate_dice_roll(no_dice, no_sims, dice_types = global$dice_names)
}
\arguments{
\item{no_sims}{number of simulations}

\item{dice_types}{names of dice types}

\item{no_dices}{list or vector indicating number of each die type to simulate}
}
\value{
vectors of rolls sums
}
\description{
Simulates dice rolls
}
