% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calculation.R
\name{simulate_die_type}
\alias{simulate_die_type}
\title{Simulates number of rolls of particular die type}
\usage{
simulate_die_type(no_dice, die_type, no_sims)
}
\arguments{
\item{no_dice}{number of dice of selected type}

\item{die_type}{string indicating die type}

\item{no_sims}{number of simulations}
}
\value{
vector of simulated roll sums
}
\description{
Simulates number of rolls of particular die type
}
