% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calculation.R
\name{simulate_iid_discrete_sum}
\alias{simulate_iid_discrete_sum}
\title{Simulates sum of a number of iid (Independent and identically distributed)
discrete random variables}
\usage{
simulate_iid_discrete_sum(no_variables, max, no_sims, min = 1)
}
\arguments{
\item{no_variables}{number of variables to simulate}

\item{max}{maximal variable value}

\item{no_sims}{number of simulations}

\item{min}{minimal variable value, default set to 1}
}
\value{
simulate sum vector
}
\description{
Simulates sum of a number of iid (Independent and identically distributed)
discrete random variables
}
