% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_init.R
\name{start_app}
\alias{start_app}
\title{Wrapper for shinyApp function to start the dashboard}
\usage{
start_app()
}
\description{
Wrapper for shinyApp function to start the dashboard
}
\examples{
\donotrun{
start_app()
}

}
