% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats_server.R
\name{sum_stats_server}
\alias{sum_stats_server}
\title{Server-side calculations for sum_stats tab}
\usage{
sum_stats_server(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\description{
Server-side calculations for sum_stats tab
}
