% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats_ui.R
\name{sum_stats_ui}
\alias{sum_stats_ui}
\title{Creates ui content for sum_stats_tab}
\usage{
sum_stats_ui()
}
\value{
list of ui elements for sum_stats_tab
}
\description{
Creates ui content for sum_stats_tab
}
